Added timer API, infinite test support and random sleep
Added timer API to make the code easier to read (and needed to stop the
timer in more places).
Added support for infinite tests (looping over the stuff defined in
test_seq...make sure that the devices ends up in the state from the
begining of the test to make sure this works.
Added SLEEP_RANDOM as an option (WIP, not sure how to configure)