Make the framework prompt aware
Now it will wait for the prompt before running the next command. This is
useful when we have commands that take a longer time to execute.
The disadvantage is that we need to tell the test which markers are
prompts and include some additional ones (like Are you sure, etc)