Improving the timeout behavior by adding more tries
Instead of timeout and that is it, try with a few CRs before. I believe
that the serial socket has some moments when it disconnects, so
overwatcher misses some output from the device. A CR should trigger
another display and get it back up and running. If it is not the case,
it just extends the dead time of the test.
To solve this I also unified the timers into one single class value.