Added support for telnet straight to device
Changes include: using a single socket, automatically re-opening the
socket when something goes wrong, changing the endline character for
telnet (added a dict to make it easier) and adding a new --telnet
parameter to the program. The scripts can be modified to work both on
serial an telnet (but need the --telnet parameter given at runtime to
know).
After these changes, overwatcher should also work with older ser2net
versions that can open only one connection to the serial port.