Improve socket handling by addinga close wrapper and other changes
Fixed an issue where the prompt was seen too many times (eol does not
need to contain #,>, as there is the timeout).
Added wrapper over socket close, with logging.
Options do not do anything if the state is the same (change needed for
telnet socket closing...sometimes it can close befor IGNORE_STATES is
called).