Add new modifier, strictStates and some cosmetic changes
Added NOSTRICT modifier - ignore state when seen.
Added the strictStates option per test. This can be used when we have
long running tests were we are doing stuff (counting, etc) and do not
care about strict order of states.
Modified the logging for the prompt wait to display command on timing
and improved the code a little.